+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Updated
 - Methods returning an array of `ZcashTransaction.Overview` try to evaluate transaction's missing blockTime. This typically applies to an expired transaction.  
 
+## Changed
+- The spelling of the `.transparent` alternative of `Entity.TransactionEntity.ZcashTransaction.Output.Pool` has been corrected from `.transaparent`.
+
 # 2.2.8 - 2025-01-10
 
 ## Added